pyproject.toml
[project]
name = "latent-spaces-by-example"
version = "0.1.0"
description = "Latent spaces by example."
readme = "README.md"
authors = [
    { name = "Sam Willis" },
    { name = "Erik Bodin" },
]
requires-python = ">=3.12"
dependencies = [
    "beartype>=0.18.5",
    "jaxtyping>=0.2.36",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
]

license = { text = "MIT" }

[project.urls]
Repository = "https://github.com/sam-willis/latent-spaces-by-example"
Issues = "https://github.com/sam-willis/latent-spaces-by-example/issues"
Documentation = "https://arxiv.org/abs/2509.23800"

[build-system]
requires = ["uv_build>=0.10.0,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "black>=26.1.0",
    "isort>=7.0.0",
    "pytest>=9.0.2",
    "ruff>=0.15.0",
]

# Optional dependencies for examples (not required for core library usage).
examples = [
    "accelerate>=0.24.0",
    "bitsandbytes>=0.47.0",
    "click>=8.1.0",
    "diffusers>=0.25.0",
    "matplotlib>=3.8.0",
    "pillow>=10.0.0",
    "protobuf>=4.25.0",
    "safetensors>=0.4.0",
    "sentencepiece>=0.2.0",
    "torch>=2.0.0",
    "tqdm>=4.66.0",
    "transformers>=4.35.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["F722"]

[tool.pytest.ini_options]
testpaths = ["tests"]
